cast to int

+++ src/att.cpp
@@ -77,7 +77,7 @@ data_VU Att::StrToVU(const std::string &value, bool logWarning) const
 std::string Att::ArticulationListToStr(data_ARTICULATION_List data) const
 {
     std::ostringstream ss;
-    for (size_t i = 0; i < data.size(); ++i) {
+    for (int i = 0; i < (int)data.size(); ++i) {
         if (i != 0) ss << " ";
         ss << ArticulationToStr(data[i]);
     }
@@ -128,7 +128,7 @@ data_BEATRPT_REND Att::StrToBeatrptRend(const std::string &value, bool logWarnin
 std::string Att::BulgeToStr(const data_BULGE &data) const
 {
     std::ostringstream ss;
-    for (size_t i = 0; i < data.size(); ++i) {
+    for (int i = 0; i < (int)data.size(); ++i) {
         if (i != 0) ss << " ";
         ss << data[i].first << " " << data[i].second;
     }
@@ -149,7 +149,7 @@ data_BULGE Att::StrToBulge(const std::string &value, bool logWarning) const
     // Convert entries to numerical values
     data_BULGE bulge;
     Att converter;
-    for (size_t i = 0; i < entries.size() - 1; i += 2) {
+    for (int i = 0; i < (int)entries.size() - 1; i += 2) {
         const double distance = converter.StrToDbl(entries[i]);
         const double offset = converter.StrToDbl(entries[i + 1]);
         if ((offset < 0.0) || (offset > 100.0)) {
@@ -375,7 +375,7 @@ std::string Att::MeasurebeatToStr(data_MEASUREBEAT data) const
 
 data_MEASUREBEAT Att::StrToMeasurebeat(std::string value, bool logWarning) const
 {
-    for (size_t i = 0; i < value.length(); ++i) {
+    for (int i = 0; i < (int)value.length(); ++i) {
         if (iswspace(value[i])) {
             value.erase(i, 1);
             i--;
@@ -666,7 +666,7 @@ data_PROLATIO Att::StrToProlatio(const std::string &value, bool logWarning) cons
 std::string Att::SummandListToStr(data_SUMMAND_List data) const
 {
     std::ostringstream ss;
-    for (size_t i = 0; i < data.size(); ++i) {
+    for (int i = 0; i < (int)data.size(); ++i) {
         if (i != 0) ss << "+";
         ss << data[i];
     }
@@ -733,7 +733,7 @@ data_TIE Att::StrToTie(const std::string &value, bool logWarning) const
 std::string Att::XsdAnyURIListToStr(xsdAnyURI_List data) const
 {
     std::ostringstream ss;
-    for (size_t i = 0; i < data.size(); ++i) {
+    for (int i = 0; i < (int)data.size(); ++i) {
         if (i != 0) ss << " ";
         ss << data[i];
     }
@@ -754,7 +754,7 @@ xsdAnyURI_List Att::StrToXsdAnyURIList(const std::string &value) const
 std::string Att::XsdPositiveIntegerListToStr(xsdPositiveInteger_List data) const
 {
     std::ostringstream ss;
-    for (size_t i = 0; i < data.size(); ++i) {
+    for (int i = 0; i < (int)data.size(); ++i) {
         if (i != 0) ss << " ";
         ss << data[i];
     }
